backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
paddlepaddle==3.2.0
paddleocr==3.3.3
torch>=2.0.0,<3
transformers>=4.40.0
pillow>=10.0.0